backend/requirements.txt
adrf==0.1.2
Django==4.2
django-filter==23.3
djangorestframework==3.14
gunicorn==21.2.0
uvicorn==0.24
